+++ lib/dead_end/api.rb
@@ -9,8 +9,6 @@ require "ripper"
 require "timeout"
 
 module DeadEnd
-  VERSION = UnloadedDeadEnd::VERSION
-
   # Used to indicate a default value that cannot
   # be confused with another input.
   DEFAULT_VALUE = Object.new.freeze